Updated enc.h
Updated enc.h to use maximum uint32 value to avoid overflow in counter. This change will also make the timer callback function for the encoder not needed anymore!

NOTE: Only TIM2 supports 32-bit values.

+++ hw/inc/enc.h
@@ -65,7 +65,7 @@
   .Init = {                                                                    \
     .Prescaler = 0,                                                            \
     .CounterMode = TIM_COUNTERMODE_UP,                                         \
-    .Period = UINT16_MAX,                                                      \
+    .Period = UINT32_MAX,                                                      \
     .RepetitionCounter = 0,                                                    \
     .ClockDivision = TIM_CLOCKDIVISION_DIV1,                                   \
     .AutoReloadPreload = TIM_AUTORELOAD_PRELOAD_DISABLE,                       \
